Extract lastUpdated logic into a hook
