+++ packages/manager/src/features/Longview/LongviewLanding/LongviewClientRow.tsx
@@ -1,6 +1,4 @@
 import Close from '@material-ui/icons/Close';
-import { APIError } from 'linode-js-sdk/lib/types';
-import { pathOr } from 'ramda';
 import * as React from 'react';
 import { compose } from 'recompose';
 
@@ -10,7 +8,7 @@ import { makeStyles, Theme } from 'src/components/core/styles';
 import Grid from 'src/components/Grid';
 import CPUGauge from './Gauges/CPU';
 
-import { getLastUpdated } from '../request';
+import { useClientLastUpdated } from '../shared/useClientLastUpdated';
 import LoadGauge from './Gauges/Load';
 import NetworkGauge from './Gauges/Network';
 import RAMGauge from './Gauges/RAM';
@@ -61,9 +59,6 @@ type CombinedProps = Props & LVDataProps & DispatchProps;
 const LongviewClientRow: React.FC<CombinedProps> = props => {
   const classes = useStyles();
 
-  let requestInterval: NodeJS.Timeout;
-  let mounted = true;
-
   const {
     clientID,
     clientLabel,
@@ -73,95 +68,10 @@ const LongviewClientRow: React.FC<CombinedProps> = props => {
     updateLongviewClient
   } = props;
 
-  /*
-   lastUpdated _might_ come back from the endpoint as 0, so it's important
-   that we differentiate between _0_ and _undefined_
-   */
-  const [lastUpdated, setLastUpdated] = React.useState<number | undefined>(
-    undefined
+  const { lastUpdated, lastUpdatedError, authed } = useClientLastUpdated(
+    clientAPIKey,
+    () => props.getClientStats(clientAPIKey)
   );
-  const currentLastUpdated = React.useRef(lastUpdated);
-  const [lastUpdatedError, setLastUpdatedError] = React.useState<
-    APIError[] | undefined
-  >();
-  const [authed, setAuthed] = React.useState<boolean>(true);
-
-  const requestAndSetLastUpdated = () => {
-    /*
-     get the current last updated value 
-
-     This function is called as a closure inside the onMount useEffect
-     so we need to use a ref to get the new value
-    */
-    const { current: newLastUpdated } = currentLastUpdated;
-
-    setLastUpdatedError(undefined);
-
-    return getLastUpdated(clientAPIKey)
-      .then(response => {
-        /*
-          only update _lastUpdated_ state if it hasn't already been set
-          or the API response is in a time past what's already been set.
-        */
-        if (
-          mounted &&
-          (typeof newLastUpdated === 'undefined' ||
-            pathOr(0, ['updated'], response) > newLastUpdated)
-        ) {
-          setLastUpdated(response.updated);
-          props.getClientStats(props.clientAPIKey);
-        }
-      })
-      .catch(e => {
-        /**
-         * The first request we make after creating a new client will almost always
-         * return an authentication failed error.
-         */
-        const reason = pathOr('', [0, 'reason'], e);
-
-        if (mounted) {
-          if (reason.match(/authentication/i)) {
-            setAuthed(false);
-          }
-
-          /* only set lastUpdated error if we haven't already gotten data before */
-          if (typeof newLastUpdated === 'undefined') {
-            setLastUpdatedError(e);
-          }
-        }
-      });
-  };
-
-  React.useEffect(() => {
-    /*
-     update the ref each time the lastUpdate state changes 
-
-     Why not just add lastUpdated as a dependency to the useEffect below?
-     Because we don't want to re-instatiate the setInterval() over and over again
-     but instead just do it once.
-
-     The closure inside the useEffect below needs to know when lastUpdated changes
-     but doesn't necessarily need to be re-defined again. useRef lets us accomplish this
-
-     See: https://github.com/facebook/react/issues/14010#issuecomment-433788147
-     
-    */
-    currentLastUpdated.current = lastUpdated;
-  }, [lastUpdated]);
-
-  /** request on first mount */
-  React.useEffect(() => {
-    requestAndSetLastUpdated().then(() => {
-      requestInterval = setInterval(() => {
-        requestAndSetLastUpdated();
-      }, 10000);
-    });
-
-    return () => {
-      mounted = false;
-      clearInterval(requestInterval);
-    };
-  }, []);
 
   /**
    * We want to show a "waiting for data" state
